% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi.R
\name{body_mass_index_classification}
\alias{body_mass_index_classification}
\title{Classify Body Mass Index}
\usage{
body_mass_index_classification(bmi, ignore = TRUE)
}
\arguments{
\item{bmi}{numeric Body Mass Index.}

\item{ignore}{boolean ignore incorrect parameter values and return NA.}
}
\value{
BMI classification
}
\description{
Classify Body Mass Index as Underweight, Normal, Overweight or Obese.
}
\examples{
body_mass_index_classification(22.9)

}
