% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdai.R
\name{cdai_classification}
\alias{cdai_classification}
\title{Calculate CDAI activity level}
\usage{
cdai_classification(cdai, ignore = TRUE)
}
\arguments{
\item{cdai}{numeric CDAI score.}

\item{ignore}{boolean ignore incorrect cdai values and return NA.}
}
\value{
CDAI level.
}
\description{
Calculate the Clinical Disease Activity Index Level for a CDAI score. Remission: >= 0 and
<= 2.8; Low: > 2.8 and <= 10; Moderate: > 10 and <= 22; High: > 22.
}
\examples{
cdai_classification(12)
cdai_classification(c(2, 42))

}
\references{
\url{http://www.ncbi.nlm.nih.gov/pubmed/16273793}
}
