% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdai.R
\name{cdai_response}
\alias{cdai_response}
\title{Calculate CDAI response}
\usage{
cdai_response(response, ignore = TRUE)
}
\arguments{
\item{response}{numeric CDAI percentage response.}

\item{ignore}{boolean ignore incorrect CDAI value types and return NA.}
}
\value{
CDAI response classification
}
\description{
Calculate the CDAI 50/70/85 for a CDAI percentage response.
}
\examples{
cdai_response(76.9)
cdai_response(c(88.0, -13.3))

}
