% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdai.R
\name{cdai_score}
\alias{cdai_score}
\title{Calculate CDAI score}
\usage{
cdai_score(tjc, sjc, ptgh, phgh, digits = 0, ignore = TRUE)
}
\arguments{
\item{tjc}{numeric Tender 28 joint count (0 to 28).}

\item{sjc}{numeric Swollen 28 joint count (0 to 28).}

\item{ptgh}{numeric Patient Global Health VAS score (0 to 10cm).}

\item{phgh}{numeric Physician Global Health VAS score (0 to 10cm).}

\item{digits}{numeric specifying the number of decimal places. Defaults to 0.}

\item{ignore}{boolean ignore incorrect parameter values and return NA.}
}
\value{
CDAI score.
}
\description{
Calculate the Clinical Disease Activity Index.
}
\examples{
cdai_score(tjc = 4, sjc = 5, ptgh = 5, phgh = 6)
cdai_score(tjc = 4, sjc = 5, ptgh = 10, phgh = 8)

}
