% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dapsa.R
\name{clinical_dapsa_score}
\alias{clinical_dapsa_score}
\title{Calculate Clinical DAPSA score}
\usage{
clinical_dapsa_score(tjc, sjc, pain, ptgh, digits = 0, ignore = TRUE)
}
\arguments{
\item{tjc}{numeric Tender 68 joint count (0 to 68).}

\item{sjc}{numeric Swollen 66 joint count (0 to 66).}

\item{pain}{numeric Pain VAS (0 to 10cm).}

\item{ptgh}{numeric Patient Global Health VAS score (0 to 10cm).}

\item{digits}{numeric specifying the number of decimal places. Defaults to 0.}

\item{ignore}{boolean ignore incorrect parameter values and return NA.}
}
\value{
Clinical DAPSA score.
}
\description{
Calculate the Clinical Disease Activity Index for Psoriatic Arthritis.
}
\examples{
clinical_dapsa_score(tjc = 4, sjc = 5, pain = 2.5, ptgh = 5)
clinical_dapsa_score(tjc = 28, sjc = 28, pain = 10, ptgh = 10)

}
