% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apr.R
\name{crp_classification}
\alias{crp_classification}
\title{Classify C-reactive protein}
\source{
Nehring SM, Goyal A, Patel BC. C Reactive Protein. 2023 Jul 10.
In: StatPearls [Internet]. Treasure Island (FL): StatPearls Publishing;
2025 Jan–. PMID: 28722873.
}
\usage{
crp_classification(crp, cutoff = NULL, ignore = TRUE)
}
\arguments{
\item{crp}{numeric C-reactive protein (mg/dL)}

\item{cutoff}{CRP cutoff for positive/negative classicication.}

\item{ignore}{boolean ignore incorrect parameter values and return NA.}
}
\value{
CRP classification - Positive or Negative
}
\description{
Classify C-reactive protein (CRP) as either positive or negative
based on CRP value.
}
\examples{
crp_classification(crp = 16)
crp_classification(crp = 12, cutoff = 10)

}
