% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dapsa.R
\name{dapsa_classification}
\alias{dapsa_classification}
\title{Calculate DAPSA activity level scores}
\usage{
dapsa_classification(dapsa, ignore = TRUE)
}
\arguments{
\item{dapsa}{numeric DAPSA score.}

\item{ignore}{boolean ignore incorrect DAPSA values and return NA.}
}
\value{
DAPSA activity level.
}
\description{
Calculate the Disease Activity Level for a DAPSA score. Remission: >= 0 and
<= 4; Low: > 4 and <= 14; Moderate: > 14 and <= 28; High: > 28.
}
\examples{
dapsa_classification(13)
dapsa_classification(266)

}
