% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dapsa.R
\name{dapsa_response}
\alias{dapsa_response}
\title{Calculate DAPSA response}
\usage{
dapsa_response(response, ignore = TRUE)
}
\arguments{
\item{response}{numeric DAPSA percentage response.}

\item{ignore}{boolean ignore incorrect DAPSA value types and return NA.}
}
\value{
DAPSA response classification
}
\description{
Calculate the DAPSA response for a DAPSA percentage response.
}
\examples{
dapsa_response(76.9)
dapsa_response(c(88.0, -13.3))

}
