% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dapsa.R
\name{dapsa_score}
\alias{dapsa_score}
\title{Calculate DAPSA score}
\usage{
dapsa_score(tjc, sjc, pain, ptgh, crp, digits = 0, crp_unit, ignore = TRUE)
}
\arguments{
\item{tjc}{numeric Tender 68 joint count (0 to 68).}

\item{sjc}{numeric Swollen 66 joint count (0 to 66).}

\item{pain}{numeric Pain VAS (0 to 10cm).}

\item{ptgh}{numeric Patient Global Health VAS score (0 to 10cm).}

\item{crp}{numeric C-reactive protein (CRP) in mg/dl.}

\item{digits}{numeric specifying the number of decimal places. Defaults to 0.}

\item{crp_unit}{character specifying either 'mg/L' or 'mg/dL'.}

\item{ignore}{boolean ignore incorrect parameter values and return NA.}
}
\value{
DAPSA score.
}
\description{
Calculate the Disease Activity Index for Psoriatic Arthritis.
}
\examples{
dapsa_score(tjc = 4, sjc = 5, pain = 2.5, ptgh = 5, crp = 7)
dapsa_score(tjc = 28, sjc = 28, pain = 10, ptgh = 10, crp = 10)

}
