% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/das.R
\name{das_28_classification}
\alias{das_28_classification}
\title{Calculate DAS activity level}
\usage{
das_28_classification(das, ignore = TRUE)
}
\arguments{
\item{das}{numeric DAS28 score.}

\item{ignore}{boolean ignore incorrect das values and return NA.}
}
\value{
DAS28 activity level.
}
\description{
Calculate the Disease Activity Level for a DAS28 score. Remission: >= 0 and
< 2.6; Low: >= 2.6 and < 3.2; Moderate: >= 3.2 and <= 5.1; High: > 5.1.
}
\examples{
das_28_classification(1.80)
das_28_classification(c(5.11, 4.25))

}
