% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{dates_to_duration}
\alias{dates_to_duration}
\title{Calculate the duration between two dates in days.}
\usage{
dates_to_duration(date1, date2)
}
\arguments{
\item{date1}{date}

\item{date2}{date}
}
\description{
Calculates how the duration in days between two dates.
}
\examples{
date1 <- as.Date("2010-01-01")
date2 <- as.Date("2010-02-13")

dates_to_duration(date1, date2)

}
