% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlqi.R
\name{dlqi_classification}
\alias{dlqi_classification}
\title{Calculate DLQI classification}
\usage{
dlqi_classification(dlqi, ignore = TRUE)
}
\arguments{
\item{dlqi}{numeric DLQI score in the range 0 to 30.}

\item{ignore}{boolean ignore incorrect dlqi scores and return NA.}
}
\value{
DLQI classification level.
}
\description{
Calculate the classification level for Dermatology Life Quality Index
}
\details{
0-1:   no effect at all on patient's life
2-5:   small effect on patient's life
6-10:  moderate effect on patient's life
11-20: very large effect on patient's life
21-30: extremely large effect on patient's life
}
\examples{
dlqi_classification(17)
dlqi_classification(c(2, 25))

}
