% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlqi.R
\name{dlqi_score}
\alias{dlqi_score}
\title{Calculate DLQI score}
\usage{
dlqi_score(q1, q2, q3, q4, q5, q6, q7, q8, q9, q10, ignore = TRUE)
}
\arguments{
\item{q1}{character Answer to DLQI question 1}

\item{q2}{character Answer to DLQI question 2}

\item{q3}{character Answer to DLQI question 3}

\item{q4}{character Answer to DLQI question 4}

\item{q5}{character Answer to DLQI question 5}

\item{q6}{character Answer to DLQI question 6}

\item{q7}{character Answer to DLQI question 7}

\item{q8}{character Answer to DLQI question 8}

\item{q9}{character Answer to DLQI question 9}

\item{q10}{character Answer to DLQI question 10}

\item{ignore}{boolean ignore incorrect dlqi parameters and return NA.}
}
\value{
DLQI classification level.
}
\description{
Calculate the Dermatology Life Quality Index score. The DLQI is a seven item
questionnaire relating to skin problems. For all questions with the exception
of question 10 the possible answers are "Very much", "A lot", "A little",
"Not at all" and "Not relevant". Question 7 is a two-part question where the
possible answers for the first part are "Yes", "No" or "Not relevant". The
second is only answered if the first part is "No". The possible answers to
the second part are "A lot", "A little" and "Not at all". Valid answers to
question 7 for this function are "Yes", "Not relevant", "A lot", "A little"
and "Not at all".
}
\examples{
dlqi_score("Very much", "A lot", "A little","Not at all", "Not relevant",
            "A lot", "Yes", "Very much", "A lot", "A little")

}
