% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apr.R
\name{esr_classification}
\alias{esr_classification}
\title{Classify Erythrocyte Sedimentation Rate}
\source{
Tishkowski K, Gupta V. Erythrocyte Sedimentation Rate. 2023 Apr 23.
In: StatPearls [Internet]. Treasure Island (FL): StatPearls Publishing;
2025 Jan–. PMID: 32491417.
}
\usage{
esr_classification(esr, age, sex, ignore = TRUE)
}
\arguments{
\item{esr}{numeric Erythrocyte Sedimentation Rate mm/hr.}

\item{age}{numeric Age in years.}

\item{sex}{character Sex - Male or Female.}

\item{ignore}{boolean ignore incorrect parameter values and return NA.}
}
\value{
ESR classification - Positive or Negative
}
\description{
Classify Erythrocyte Sedimentation Rate (ESR) as either positive or negative
based on ESR, age and sex.
}
\examples{
esr_classification(esr = 16, age = 25, sex = "Male")
esr_classification(esr = 10, age = 55, sex = "female")

}
