% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/das.R
\name{eular_ra_response}
\alias{eular_ra_response}
\title{Calculate EULAR response}
\usage{
eular_ra_response(das1, das2, ignore = TRUE)
}
\arguments{
\item{das1}{numeric first DAS28 score}

\item{das2}{numeric follow-up DAS28 score}

\item{ignore}{boolean ignore incorrect das values and return NA.}
}
\value{
EULAR response classification
}
\description{
Calculate the EULAR response between two DAS28 scores.
}
\examples{
eular_ra_response(das1 = 5.21, das2 = 1.8)
eular_ra_response(das1 = c(4.3, 3.93), das2 = c(5.13, 1.1))

}
