% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_date}
\alias{is_date}
\title{Test for objects of type "Date"}
\usage{
is_date(x)
}
\arguments{
\item{x}{object to be tested.}
}
\description{
Checks if an object inherits class Date. Used prior to calculating durations
from two dates.
}
\examples{
date <- as.Date("2010-01-01")

is_date(date)

}
