% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{percentage_response}
\alias{percentage_response}
\title{Calculate percentage response}
\usage{
percentage_response(score1, score2, ignore = TRUE)
}
\arguments{
\item{score1}{numeric First score.}

\item{score2}{numeric Second (more recent) score.}

\item{ignore}{boolean ignore incorrect values and return NA.}
}
\value{
Percentage response
}
\description{
Calculate the percentage response between two activity scores.
}
\examples{
percentage_response(score1 = 13, score2 = 3)
percentage_response(score1 = c(266, 150), score2 = c(32, 170))

}
