% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psaid.R
\name{psaid_12_score}
\alias{psaid_12_score}
\title{Calculate PSAID-12 score}
\usage{
psaid_12_score(
  pain,
  fatigue,
  skin,
  work,
  functional,
  discomfort,
  sleep,
  coping,
  anxiety,
  shame,
  social,
  depression,
  digits = 1,
  ignore = TRUE
)
}
\arguments{
\item{pain}{numeric Pain due PsA during the last week (0: none to 10: extreme).}

\item{fatigue}{numeric. Level of the fatigue due PsA during the last week (0: no fatigue to 10: totally exhausted).}

\item{skin}{numeric Skin problems due PsA during the last week (0: none to 10: extreme).}

\item{work}{numeric Difficulties participating fully in work and/or leisure activities due to PsA during the last week (0: none to 10: extreme).}

\item{functional}{numeric. Difficulty doing daily physical activities due PsA during the last week (0: no difficulty to 10: extreme difficulty).}

\item{discomfort}{numeric. Feeling of discomfort and annoyance with everyday tasks due to PsA during the last week (0: none to 10: extreme).}

\item{sleep}{numeric. Sleep difficulties (i.e. resting at night) due to PsA during the last week (0: no difficulty to 10: extreme difficulty).}

\item{coping}{numeric. Coping (manage, deal, make do) with PsA during the last week (0: very well to 10: very poorly).}

\item{anxiety}{numeric. Level of anxiety, fear and uncertainty due to your PsA during the last week (0: none to 10: extreme).}

\item{shame}{numeric. Level of embarrassment and/or shame due to appearance experienced during the last week (0: none to 10: extreme).}

\item{social}{numeric. Difficulties participating fully in social activities due to PsA during the last week (0: none to 10: extreme).}

\item{depression}{numeric. Level of depression due to PsA experienced during the last week (0: none to 10: extreme).}

\item{digits}{numeric specifying the number of decimal places. Defaults to 1.}

\item{ignore}{boolean ignore incorrect parameter values and return NA.}
}
\value{
PSAID-12 score.
}
\description{
Calculate the EULAR Psoriatic Arthritis Impact of Disease 12 item score.
}
\examples{
psaid_12_score(8, 10, 6, 6, 6, 9, 9, 5, 10, 8, 6, 1)

}
