% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdai.R
\name{sdai_classification}
\alias{sdai_classification}
\title{Calculate SDAI activity level}
\usage{
sdai_classification(sdai, ignore = TRUE)
}
\arguments{
\item{sdai}{numeric SDAI score.}

\item{ignore}{boolean ignore incorrect sdai values and return NA.}
}
\value{
SDAI level.
}
\description{
Calculate the Simple Disease Activity Index Level for a SDAI score. Remission: >= 0 and
<= 3.3; Low: > 3.3 and <= 11; Moderate: > 11 and <= 26; High: > 26.
}
\examples{
sdai_classification(12)
sdai_classification(c(2, 42))

}
