% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdai.R
\name{sdai_score}
\alias{sdai_score}
\title{Calculate SDAI score}
\usage{
sdai_score(tjc, sjc, ptgh, phgh, crp, digits = 0, ignore = TRUE)
}
\arguments{
\item{tjc}{numeric Tender 28 joint count (0 to 28).}

\item{sjc}{numeric Swollen 28 joint count (0 to 28).}

\item{ptgh}{numeric Patient Global Health VAS score (0 to 10cm).}

\item{phgh}{numeric Physician Global Health VAS score (0 to 10cm).}

\item{crp}{numeric C-reactive protein (CRP) in mg/dl.}

\item{digits}{numeric specifying the number of decimal places. Defaults to 0.}

\item{ignore}{boolean ignore incorrect parameter values and return NA.}
}
\value{
SDAI score.
}
\description{
Calculate the Simple Disease Activity Index.
}
\examples{
sdai_score(tjc = 4, sjc = 5, ptgh = 5, phgh = 6, crp=7)
sdai_score(tjc = 4, sjc = 5, ptgh = 10, phgh = 8, crp=3)

}
\references{
\url{http://www.ncbi.nlm.nih.gov/pubmed/16273793}
}
